Branch erstellt

+++ include/lvr2/reconstruction/AdaptiveKSearchSurface.tcc
@@ -183,7 +183,8 @@ void AdaptiveKSearchSurface<BaseVecT>::calculateSurfaceNormals()
         // search on the stann kd tree. So we don't use
         // the template parameter T for di
         vector<size_t> id;
-        vector<float> di;
+        //WAS FLOAT BEFORE
+        vector<double> di;
 
         int n = 0;
         size_t k = k_0;
@@ -339,7 +340,8 @@ void AdaptiveKSearchSurface<BaseVecT>::interpolateSurfaceNormals()
     for( int i = 0; i < (int)numPoints; i++)
     {
         vector<size_t> id;
-        vector<float> di;
+        //WAS FLOAZ BEFORE
+        vector<double> di;
 
         this->m_searchTree->kSearch(pts[i], this->m_ki, id, di);
 
@@ -464,7 +466,8 @@ pair<typename BaseVecT::CoordType, typename BaseVecT::CoordType>
     int k = this->m_kd;
 
     vector<size_t> id;
-    vector<float> di;
+    //Was FLOAT BEFORE
+    vector<double> di;
 
     //Allocate ANN point
     {
